requirements.txt
Flask
newspaper3k
google-generativeai
python-dotenv
Flask-CORS
requests
